Refactor CoseKey and CoseKeyPrivate to use CoseEcCurve enum instead of ECCurveType
Refactor the CoseKey and CoseKeyPrivate structs in the CoseKey.swift file to use the CoseEcCurve enum instead of the ECCurveType enum. This change ensures consistency and clarity in the codebase.

Refactor DeviceEngagement to use CoseEcCurve enum in initializer

Refactor the DeviceEngagement struct in the DeviceEngagement.swift file to use the CoseEcCurve enum in the initializer instead of the ECCurveType enum. This change improves code readability and maintainability.

Update CoseEcCurve to use UInt64 raw value type

Update the CoseEcCurve enum in the CoseEcCurve.swift file to use UInt64 as the raw value type instead of Int. This change aligns with the COSE registry and ensures compatibility with iOS natively supported curves.

Refactor KeyOptions.swift to use requireApplicationPassword instead of requirePassword

Refactor the KeyOptions.swift file to use the requireApplicationPassword property instead of the requirePassword property in the KeyAccessControl struct. This change improves clarity and consistency in the codebase.

+++ Sources/MdocDataModel18013/SecureArea/CoseEcCurve.swift
@@ -16,8 +16,10 @@ limitations under the License.
 
 import Foundation
 
- /// Elliptic curve identifiers  from the [IANA COSE registry](https://www.iana.org/assignments/cose/cose.xhtml).
-public enum CoseEcCurve: Int, Sendable {
+/// Elliptic curve identifiers  from the [IANA COSE registry](https://www.iana.org/assignments/cose/cose.xhtml).
+
+/// This list is a superset of the iOS natively supported curves
+public enum CoseEcCurve: UInt64, Sendable {
     /// The curve identifier for P-256
     case P256 = 1
     /// The curve identifier for P-384
@@ -69,7 +71,7 @@ public enum CoseEcCurve: Int, Sendable {
         "brainpoolP512r1": .BRAINPOOLP512R1
     ]
 
-    static func fromInt(_ coseCurveIdentifier: Int) throws -> CoseEcCurve {
+    static func fromInt(_ coseCurveIdentifier: UInt64) throws -> CoseEcCurve {
         guard let curve = CoseEcCurve(rawValue: coseCurveIdentifier) else {
             throw NSError(domain: "EcCurve", code: -1, userInfo: [NSLocalizedDescriptionKey: "No curve with COSE identifier \(coseCurveIdentifier)"])
         }
